Move scheduled jobs under a dedicated folder
