+++ src/seeds.php
@@ -18,21 +18,18 @@ if ($environment === 'development') {
 
 $demo = \Minz\Configuration::$application['demo'];
 if ($demo) {
-    $reset_demo_job = new \flusio\jobs\ResetDemo();
+    $reset_demo_job = new \flusio\jobs\scheduled\ResetDemo();
     $job_dao = new \flusio\models\dao\Job();
     if (!$job_dao->findBy(['name' => $reset_demo_job->name])) {
-        $reset_demo_job->perform_at = \Minz\Time::relative('tomorrow 2:00');
-        $reset_demo_job->frequency = '+1 day';
         $reset_demo_job->performLater();
     }
 }
 
 $subscriptions_enabled = \Minz\Configuration::$application['subscriptions_enabled'];
 if ($subscriptions_enabled) {
-    $subscriptions_sync_job = new \flusio\jobs\SubscriptionsSync();
+    $subscriptions_sync_job = new \flusio\jobs\scheduled\SubscriptionsSync();
     $job_dao = new \flusio\models\dao\Job();
     if (!$job_dao->findBy(['name' => $subscriptions_sync_job->name])) {
-        $subscriptions_sync_job->frequency = '+4 hours';
         $subscriptions_sync_job->performLater();
     }
 }
